debt: playground: extract notifications-ui + simulation from main.js
Final two extractions that bring main.js under the 600-line cap.

- main/notifications-ui.js: wireNotificationUI + flashTestBtn. DOM
  wiring for the install button, push-notification toggle, per-
  category checkboxes, and per-category test buttons. All the
  notifications-module imports that only this UI needs moved here
  too; main.js keeps just captureInstallPrompt + initNotifications.
- main/simulation.js: setupFAB, togglePlay, updateFABIcon,
  formatTimeOfDay (previously exported from main.js), getTimeOfDay,
  simLoop. Owns lastFrame + simTimeAccum. Reads running via ESM
  live binding from main.js and writes through setRunning.
  Exposes resetSimulationTime() so main.js's resetSim can reset the
  accumulator.

main.js exports added: setRunning, simSpeed (read by simulation.js
simLoop); logs.js updated to import formatTimeOfDay from
simulation.js instead of main.js.

main.js: 699 → 505 lines. Under cap. npm test:unit 782/782 pass.
knip clean.
